Updated software architecture of system
