+++ config/players.json
@@ -1,0 +1,7 @@
+[
+  {
+    "id": 1,
+    "name": "aaaaaaaaa",
+    "points": 0
+  }
+]
