+++ constants.py
@@ -1,0 +1,4 @@
+
+QUESTIONS_PATH = 'config/questions.json'
+PLAYERS_PATH = 'config/players.json'
+STATE_PATH = 'state/state.json'
