--- players.json
@@ -1,17 +1,0 @@
-[
-  {
-    "id": 1,
-    "name": "Jansky + Vincent",
-    "points": 5600
-  },
-  {
-    "id": 2,
-    "name": "David + Sigal",
-    "points": 5900
-  },
-  {
-    "id": 3,
-    "name": "Marlon + Nik",
-    "points": 2101
-  }
-]
