--- state.json
@@ -1,44 +1,0 @@
-{
-  "round": 2,
-  "categories1": [
-    "Indie",
-    "Abkürzungen",
-    "Was ist?",
-    "Kamehameha",
-    "Zitate"
-  ],
-  "categories2": [
-    "LoL eSports",
-    "AI Goehte",
-    "Werbung",
-    "Nani?",
-    "Trivia"
-  ],
-  "answered": [
-    38,
-    4,
-    14,
-    7,
-    3,
-    5,
-    39,
-    48,
-    15,
-    2,
-    49,
-    50,
-    40,
-    10,
-    47,
-    9,
-    13,
-    1,
-    37,
-    8,
-    46,
-    12,
-    11,
-    36,
-    6
-  ]
-}
