+++ state/state.json
@@ -1,0 +1,18 @@
+{
+  "round": 2,
+  "categories1": [
+    "Squad Members",
+    "HUD",
+    "Sprich Deutsch du Hurensohn",
+    "Trivia",
+    "Pixel Cover"
+  ],
+  "categories2": [
+    "Google Translate",
+    "GeoGamer",
+    "Item",
+    "Komisches Sigal Spiel",
+    "OST"
+  ],
+  "answered": []
+}
